network-routing-system-v24/requirements.txt
pyyaml>=6.0
psutil>=5.9.0
requests>=2.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
